Refactor NoSuchSubscriptionException
