--- src/main/java/org/zalando/nakadi/exceptions/NoSuchSubscriptionException.java
@@ -1,19 +1,0 @@
-package org.zalando.nakadi.exceptions;
-
-import javax.ws.rs.core.Response;
-
-public class NoSuchSubscriptionException extends NakadiException {
-
-    public NoSuchSubscriptionException(final String message) {
-        super(message);
-    }
-
-    public NoSuchSubscriptionException(final String msg, final Exception cause) {
-        super(msg, cause);
-    }
-
-    @Override
-    protected Response.StatusType getStatus() {
-        return Response.Status.NOT_FOUND;
-    }
-}
