+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoSuchSubscriptionException.java
@@ -1,0 +1,12 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class NoSuchSubscriptionException extends NakadiRuntimeBaseException {
+
+    public NoSuchSubscriptionException(final String message) {
+        super(message);
+    }
+
+    public NoSuchSubscriptionException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+}
